Add files via upload
Code I created for an open-ended Arduino project in my Computer Organization class. I decided to attempt creating a game utilizing buttons and an LCD display that I wired to the Arduino. For the game itself, I made a moveable character that needs to avoid standing on the same display segment as a block that spawns in or else the game is lost.